backend/requirements.txt
Flask
torch
torchvision
opencv-python
pillow
transformers
flask-cors
pandas
requests
